add(server): github oauth & routes jwt auth

+++ server/src/routes/memories.ts
@@ -1,78 +1,104 @@
-import { FastifyInstance } from "fastify";
+import { FastifyInstance } from 'fastify'
 
-import { z } from "zod";
-import { prisma } from "../lib/prisma";
+import { z } from 'zod'
+import { prisma } from '../lib/prisma'
 
 export async function memoriesRoutes(app: FastifyInstance) {
-  app.get("/memories", async () => {
+  app.addHook('preHandler', async (request) => {
+    // preHandler hook runs before the route handler for ALL routes
+    await request.jwtVerify() // verify the JWT token before proceeding
+  })
+
+  app.get('/memories', async (request) => {
     const memories = await prisma.memory.findMany({
-      orderBy: {
-        createdAt: "asc",
+      where: {
+        userId: request.user.sub,
       },
-    });
+      orderBy: {
+        createdAt: 'asc',
+      },
+    })
 
     return memories.map((memory) => {
       return {
         id: memory.id,
         coverUrl: memory.coverUrl,
-        excerpt: memory.content.substring(0, 115).concat("..."),
-      };
-    });
-  });
+        excerpt: memory.content.substring(0, 115).concat('...'),
+      }
+    })
+  })
 
-  app.get("/memories/:id", async (request) => {
+  app.get('/memories/:id', async (request, reply) => {
     const paramsSchema = z.object({
       id: z.string().uuid(),
-    });
+    })
 
-    const { id } = paramsSchema.parse(request.params);
+    const { id } = paramsSchema.parse(request.params)
 
     const memory = await prisma.memory.findUniqueOrThrow({
       where: {
         id,
       },
-    });
+    })
 
-    return memory;
-  });
+    if (!memory.isPublic && memory.userId !== request.user.sub) {
+      return reply.status(401).send({
+        message: 'You are not authorized to access this memory',
+      })
+    }
 
-  app.post("/memories", async (request) => {
+    return memory
+  })
+
+  app.post('/memories', async (request) => {
     const bodySchema = z.object({
       content: z.string(),
       coverUrl: z.string(),
       isPublic: z.coerce.boolean().default(false),
-    });
+    })
 
-    const { content, coverUrl, isPublic } = bodySchema.parse(request.body);
+    const { content, coverUrl, isPublic } = bodySchema.parse(request.body)
 
     const memory = await prisma.memory.create({
       data: {
         content,
         coverUrl,
         isPublic,
-        userId: "877182fc-a69a-48a5-9f36-786dc5fc4e25",
+        userId: request.user.sub,
       },
-    });
+    })
 
-    return memory;
-  });
+    return memory
+  })
 
-  app.put("/memories/:id", async (request) => {
+  app.put('/memories/:id', async (request, reply) => {
     const paramsSchema = z.object({
       id: z.string().uuid(),
-    });
+    })
 
-    const { id } = paramsSchema.parse(request.params);
+    const { id } = paramsSchema.parse(request.params)
 
     const bodySchema = z.object({
       content: z.string(),
       coverUrl: z.string(),
       isPublic: z.coerce.boolean().default(false),
-    });
+    })
 
-    const { content, coverUrl, isPublic } = bodySchema.parse(request.body);
+    const { content, coverUrl, isPublic } = bodySchema.parse(request.body)
 
-    const memory = await prisma.memory.update({
+    let memory = await prisma.memory.findUniqueOrThrow({
+      where: {
+        id,
+      },
+    })
+
+    if (!memory.isPublic && memory.userId !== request.user.sub) {
+      return reply.status(401).send({
+        message: 'You are not authorized to update this memory',
+      })
+    }
+
+    memory = await prisma.memory.update({
       where: {
         id,
       },
@@ -81,22 +107,34 @@ export async function memoriesRoutes(app: FastifyInstance) {
         coverUrl,
         isPublic,
       },
-    });
+    })
 
-    return memory;
-  });
+    return memory
+  })
 
-  app.delete("/memories/:id", async (request) => {
+  app.delete('/memories/:id', async (request, reply) => {
     const paramsSchema = z.object({
       id: z.string().uuid(),
-    });
+    })
 
-    const { id } = paramsSchema.parse(request.params);
+    const { id } = paramsSchema.parse(request.params)
+
+    const memory = await prisma.memory.findUniqueOrThrow({
+      where: {
+        id,
+      },
+    })
+
+    if (!memory.isPublic && memory.userId !== request.user.sub) {
+      return reply.status(401).send({
+        message: 'You are not authorized to update this memory',
+      })
+    }
 
     await prisma.memory.delete({
       where: {
         id,
       },
-    });
-  });
+    })
+  })
 }
